Renamed action modals module.
