+++ src/js/rx/containers/History.jsx
@@ -5,7 +5,7 @@ import _ from 'lodash';
 
 import SortableTable from '../../common/components/SortableTable';
 import { loadPrescriptions } from '../actions/prescriptions';
-import { openGlossaryModal } from '../actions/modal';
+import { openGlossaryModal } from '../actions/modals';
 import Pagination from '../../common/components/Pagination';
 import SortMenu from '../components/SortMenu';
 import { glossary, rxStatuses } from '../config.js';
